#965-filesystem-based-schema-registry: Some refactoring
