+++ factcast-store-pgsql/src/main/java/org/factcast/store/pgsql/registry/classpath/ClasspathRegistryFileFetcher.java
@@ -16,64 +16,22 @@
 package org.factcast.store.pgsql.registry.classpath;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 
-import org.factcast.store.pgsql.registry.RegistryFileFetcher;
-import org.factcast.store.pgsql.registry.SchemaRegistryUnavailableException;
-import org.factcast.store.pgsql.registry.transformation.TransformationSource;
-import org.factcast.store.pgsql.registry.validation.schema.SchemaSource;
+import org.factcast.store.pgsql.registry.AbstractRegistryFileFetcher;
 import org.springframework.core.io.ClassPathResource;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class ClasspathRegistryFileFetcher implements RegistryFileFetcher {
+public class ClasspathRegistryFileFetcher extends AbstractRegistryFileFetcher {
 
     private final @NonNull String base;
 
     @Override
-    public String fetchTransformation(@NonNull TransformationSource key) {
-        String path = base + "/" + key.ns() + "/" + key.type() + "/" + key.from() + "-" + key.to()
-                + "/transform.js";
-        return fetch(path);
-
-    }
-
-    private @NonNull String fetch(String path) {
-        try {
-            File file = new ClassPathResource(path).getFile();
-            if (file.exists()) {
-                return readFile(file);
-            } else {
-                throw new SchemaRegistryUnavailableException(
-                        new FileNotFoundException("Resource "
-                                + path
-                                + " does not exist."));
-            }
-        } catch (IOException e) {
-            throw new SchemaRegistryUnavailableException(e);
-        }
-    }
-
-    private static @NonNull String readFile(@NonNull File file)
-            throws IOException {
-        StringBuilder sb = new StringBuilder();
-        java.nio.file.Files.lines(file.toPath()).forEachOrdered(l -> {
-            if (sb.length() > 0) {
-                sb.append("\n");
-            }
-            sb.append(l);
-        });
-        return sb.toString();
-    }
-
-    @Override
-    public String fetchSchema(@NonNull SchemaSource key) {
-        String path = base + "/" + key.ns() + "/" + key.type() + "/" + key.version()
-                + "/schema.json";
-        return fetch(path);
+    protected File getFile(String subPath) throws IOException {
+        return new ClassPathResource(base + "/" + subPath).getFile();
     }
 
 }
